+++ src/main/java/com/shape/web/controller/ProjectController.java
@@ -59,8 +59,8 @@ public class ProjectController {
     */
     @RequestMapping(value = "/room", method = RequestMethod.POST)    //프로젝트 개설
     public String MakeRoom(@RequestParam(value = "name") String name, Authentication authentication) {
-        User user=us.getById(authentication.getName()); //유저 객체 반환
-        Integer userIdx=user.getUseridx();
+        User user = us.getById(authentication.getName()); //유저 객체 반환
+        Integer userIdx = user.getUseridx();
         Project project = new Project(name, userIdx, "");
         us.addProject(userIdx, project);
         return "redirect:/projectmanager";
@@ -78,10 +78,11 @@ public class ProjectController {
     @RequestMapping(value = "/room/{projectIdx}", method = RequestMethod.PUT)    //프로젝트 삭제
     @ResponseBody
     public void updadeRoom(@PathVariable("projectIdx") Integer projectIdx) {
-        Project project= pjs.get(projectIdx);
+        Project project = pjs.get(projectIdx);
         project.setProcessed(false);
         pjs.save(project);
     }
+
     /*
        To find out invited user
        */
@@ -114,7 +115,7 @@ public class ProjectController {
                                @RequestParam("projectIdx") Integer projectIdx,
                                Authentication authentication) {
         logger.info("Invite Member");
-        User actor=us.getById(authentication.getName()); //초대한 사람
+        User actor = us.getById(authentication.getName()); //초대한 사람
         User user = us.getById(userId); // 초대받은 사람
         Project project = pjs.get(projectIdx); // 초대받은 프로젝트
         Alarm alarm = new Alarm(0, null, null, new Date());
@@ -123,24 +124,6 @@ public class ProjectController {
         alarm.setProject(project);
         as.save(alarm); //알람 생성
         return String.valueOf(user.getUseridx());
-    }
-
-
-    /*
-    To make to-do list
-    */
-    @RequestMapping(value = "/Todolist", method = RequestMethod.POST)
-    @ResponseBody
-    public String makeTodolist(@RequestParam("projectIdx") Integer projectIdx,
-                               @RequestParam("userId") String userId,
-                               @ModelAttribute("todolist") Todolist todolist) {
-        Project project = pjs.get(projectIdx);
-        User user = us.getById(userId);
-        todolist.setProject(project); // todolist가 어디 프로젝트에서 생성되었는가
-        todolist.setUser(user); // todolist가 누구것인가
-        ts.save(todolist); // todolist 생성
-        logger.info("todolist 만듬");
-        return "Ok";
     }
